* Fixed warnings in MADARA compilation on VS 2017

+++ tests/test_circular_buffer.cpp
@@ -249,7 +249,7 @@ void test_container()
   TEST_EQ(buf.remaining(), 8UL);
   TEST_EQ(buf.count(), 10UL);
 
-  int c = buf.get_record().get_history_oldest_index();
+  int c = (int)buf.get_record().get_history_oldest_index();
   for (const auto& cur : buf.get_record().get_history())
   {
     std::cerr << c << ": " << cur << std::endl;
